Move the validator to service_drivers dir as Nick suggested
